+++ src/main.rs
@@ -2,13 +2,14 @@ use anyhow::Result;
 use yansi::Paint;
 
 use genact::args::parse_args;
-use genact::{exit_handler, run, ALL_MODULES};
+use genact::{exit_handler, run, ALL_MODULES, SPEED_FACTOR};
 
 #[async_std::main]
 async fn main() -> Result<()> {
     Paint::enable_windows_ascii();
 
     let appconfig = parse_args();
+    *SPEED_FACTOR.lock().await = appconfig.speed_factor;
 
     if appconfig.list_modules_and_exit {
         println!("Available modules:");